requirements.txt
python-dotenv
flask
streamlit

langchain
langchain-community 
langchain-core
tiktoken
openai==1.55.3
faiss-cpu

serpapi
newspaper3k
lxml_html_clean
requests
